requirements.txt
torch>=1.11.0
torchvision>=0.12.0
ultralytics>=8.0.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
matplotlib>=3.2.2
seaborn>=0.11.0
gitpython>=3.1.30